+++ src/main/java/oit/is/offline/jinrou/model/RandomRole.java
@@ -1,7 +1,6 @@
 package oit.is.offline.jinrou.model;
 
 import java.util.Random;
-import java.util.ArrayList;
 
 public class RandomRole{
   Random random = new Random();
