+++ src/main/java/oit/is/offline/jinrou/model/Vote.java
@@ -1,36 +1,36 @@
 package oit.is.offline.jinrou.model;
 
-public class Vote{
+public class Vote {
 
-  public int Voting(int num, int[] count){
-    int vc = 0;   //投票した人数 num->生きている人数
-    int s = 0;  //maxの時のユーザー番号
-    int es= 0;
+  public int Voting(int num, int[] count) {
+    int vc = 0; // 投票した人数 num->生きている人数
+    int s = 0; // maxの時のユーザー番号
+    int es = 0;
 
-    for(int i = 0; i< num; i++){
+    for (int i = 0; i < 10; i++) {
       vc += count[i];
     }
 
-    if(vc == num){
+    if (vc == num) {
       int max = count[0];
 
-      for(int j = 0; j < num; j++){ //maxを見つける
-        if(count[j] >= max){
+      for (int j = 0; j < 10; j++) { // maxを見つける
+        if (count[j] >= max) {
           max = count[j];
-          s = j+1;
+          s = j + 1;
         }
       }
 
-      for(int j = 0; j <num; j++){  //maxと同じ添え字はあるか
-        if(count[j] == max){
+      for (int j = 0; j < 10; j++) { // maxと同じ添え字はあるか
+        if (count[j] == max) {
           es++;
         }
       }
 
-      if(es ==1){
+      if (es == 1) {
         return s;
       }
-      if(es != 1){
+      if (es != 1) {
         return -1;
       }
     }
