+++ src/main/resources/templates/night.html
@@ -126,6 +126,46 @@
         <a href="/morning">霊媒を終了し、朝のページに移動する</a>
       </div>
 
+      <div th:if="${rolename eq '騎士'}">
+        <div>あなたは「騎士」です</div>
+        <div th:if="${flag3 eq 1}">
+          <div>守りたいユーザーを選択してください</div>
+          <p th:if="${user1}">
+            <a href="/knight/user1">[[${user1}]]</a>
+          </p>
+          <p th:if="${user2}">
+            <a href="/knight/user2">[[${user2}]]</a>
+          </p>
+          <p th:if="${user3}">
+            <a href="/knight/user3">[[${user3}]]</a>
+          </p>
+          <p th:if="${user4}">
+            <a href="/knight/user4">[[${user4}]]</a>
+          </p>
+          <p th:if="${user5}">
+            <a href="/knight/user5">[[${user5}]]</a>
+          </p>
+          <p th:if="${user6}">
+            <a href="/knight/user6">[[${user6}]]</a>
+          </p>
+          <p th:if="${user7}">
+            <a href="/knight/user7">[[${user7}]]</a>
+          </p>
+          <p th:if="${user8}">
+            <a href="/knight/user8">[[${user8}]]</a>
+          </p>
+          <p th:if="${user9}">
+            <a href="/knight/user9">[[${user9}]]</a>
+          </p>
+          <p th:if="${user10}">
+            <a href="/knight/user10">[[${user10}]]</a>
+          </p>
+        </div>
+      </div>
+      <div th:if="${knight}">
+        今夜は[[${knight}]]を守ります。</br>
+        <a href="/morning">騎士の夜の行動を終了し、朝のページに移動する</a>
+      </div>
       <div th:if="${rolename eq '人狼'}">
         <div>あなたは「人狼」です</div>
         <div th:if="${flag4 eq 1}">
@@ -177,46 +217,6 @@
       <div th:if="${at}">
         今夜は[[${at}]]を襲撃します。</br>
         <a href="/morning">人狼の夜の行動を終了し、朝のページに移動する</a>
-      </div>
-      <div th:if="${rolename eq '騎士'}">
-        <div>あなたは「騎士」です</div>
-        <div th:if="${flag3 eq 1}">
-          <div>守りたいユーザーを選択してください</div>
-          <p th:if="${user1}">
-            <a href="/knight/user1">[[${user1}]]</a>
-          </p>
-          <p th:if="${user2}">
-            <a href="/knight/user2">[[${user2}]]</a>
-          </p>
-          <p th:if="${user3}">
-            <a href="/knight/user3">[[${user3}]]</a>
-          </p>
-          <p th:if="${user4}">
-            <a href="/knight/user4">[[${user4}]]</a>
-          </p>
-          <p th:if="${user5}">
-            <a href="/knight/user5">[[${user5}]]</a>
-          </p>
-          <p th:if="${user6}">
-            <a href="/knight/user6">[[${user6}]]</a>
-          </p>
-          <p th:if="${user7}">
-            <a href="/knight/user7">[[${user7}]]</a>
-          </p>
-          <p th:if="${user8}">
-            <a href="/knight/user8">[[${user8}]]</a>
-          </p>
-          <p th:if="${user9}">
-            <a href="/knight/user9">[[${user9}]]</a>
-          </p>
-          <p th:if="${user10}">
-            <a href="/knight/user10">[[${user10}]]</a>
-          </p>
-        </div>
-      </div>
-      <div th:if="${knight}">
-        今夜は[[${knight}]]を守ります。</br>
-        <a href="/morning">騎士の夜の行動を終了し、朝のページに移動する</a>
       </div>
     </div>
   </div>
