+++ src/main/resources/templates/vote.html
@@ -22,8 +22,8 @@
   </h1>
 
   <h2>
-    <p th:if="${death}">
-      投票により[[${death}]]が選ばれました。</br>
+    <p th:if="${voted}">
+      投票により[[${voted}]]が選ばれました。</br>
       <a href="/night">
         夜に移行する
       </a>
